api/requirements.txt
fastapi[standard,docs,dev]
Jinja2
numpy
torch
torchvision
opencv-python
pillow 
python-multipart